Fix clippy

+++ crates/iota-core/src/authority_server.rs
@@ -868,7 +868,7 @@ impl ValidatorService {
 
         self.metrics
             .handle_soft_bundle_certificates_size_bytes
-            .observe(total_size_bytes as u64);
+            .observe(total_size_bytes);
 
         // Now that individual certificates are valid, we check if the bundle is valid.
         self.soft_bundle_validity_check(&certificates, &epoch_store, total_size_bytes)
